Use new status command
