+++ caldir-cli/src/commands/mod.rs
@@ -2,8 +2,7 @@
 // pub mod new;
 // pub mod pull;
 // pub mod push;
-// pub mod status;
-pub mod status2;
+pub mod status;
 
 /// Number of days to sync in each direction (past and future)
 pub const SYNC_DAYS: i64 = 365;
