+++ caldir-cli/src/commands/status.rs
@@ -1,96 +1,37 @@
-use std::path::PathBuf;
-
 use anyhow::Result;
+use indicatif::{ProgressBar, ProgressStyle};
 
-use crate::{config, diff};
+use crate::caldir::Caldir;
 
-use super::{CalendarContext, require_calendars};
+pub async fn run() -> Result<()> {
+    let caldir = Caldir::load()?;
+    let calendars = caldir.calendars();
 
-struct Provider(String);
+    for (i, cal) in calendars.iter().enumerate() {
+        // Show spinner while diff is loading:
+        let spinner = ProgressBar::new_spinner();
+        spinner.set_style(
+            ProgressStyle::default_spinner()
+                .tick_strings(&["-", "\\", "|", "/"])
+                .template("{msg} {spinner}")
+                .unwrap(),
+        );
+        spinner.set_message(cal.render());
+        spinner.enable_steady_tick(std::time::Duration::from_millis(80));
 
-struct Calendar {
-    path: PathBuf,
-    provider: Provider,
-}
+        let diff = cal.get_diff().await?;
 
-pub async fn run(verbose: bool) -> Result<()> {
-    let cfg = config::load_config()?;
-    require_calendars(&cfg)?;
+        spinner.finish_and_clear();
 
-    let mut any_changes = false;
+        // Finished loading, show calendar + diff:
+        println!("{}", cal.render());
+        println!("{}", diff.render());
 
-    for (calendar_name, calendar_config) in &cfg.calendars {
-        let ctx = CalendarContext::load(&cfg, calendar_name, calendar_config, verbose).await?;
-
-        if !ctx.sync_diff.has_pull_changes() && !ctx.sync_diff.has_push_changes() {
-            continue;
+        // Add spacing between calendars (but not after the last one)
+        if i < calendars.len() - 1 {
+            println!();
         }
-
-        any_changes = true;
-        print_diff(&ctx.metadata.calendar_name, &ctx.sync_diff, verbose);
-    }
-
-    if !any_changes {
-        println!("Everything up to date.");
-    } else {
-        println!("\nRun `caldir-cli pull` to pull changes, or `caldir-cli push` to push changes.");
     }
 
     Ok(())
-}
-
-fn print_diff(calendar_name: &str, sync_diff: &diff::SyncDiff, verbose: bool) {
-    println!("\n📅 {}", calendar_name);
-
-    if sync_diff.has_pull_changes() {
-        println!("  To pull:");
-        for change in &sync_diff.to_pull_create {
-            println!("    + {}", change.filename);
-        }
-        for change in &sync_diff.to_pull_update {
-            println!("    ~ {}", change.filename);
-            print_property_changes(change, verbose);
-        }
-        for change in &sync_diff.to_pull_delete {
-            println!("    - {}", change.filename);
-        }
-    }
-
-    if sync_diff.has_push_changes() {
-        println!("  To push:");
-        for change in &sync_diff.to_push_create {
-            println!("    + {}", change.filename);
-        }
-        for change in &sync_diff.to_push_update {
-            println!("    ~ {}", change.filename);
-            print_property_changes(change, verbose);
-        }
-        for change in &sync_diff.to_push_delete {
-            println!("    - {} (delete from remote)", change.uid);
-        }
-    }
-}
-
-fn print_property_changes(change: &diff::SyncChange, verbose: bool) {
-    if !verbose || change.property_changes.is_empty() {
-        return;
-    }
-
-    for prop_change in &change.property_changes {
-        match (&prop_change.old_value, &prop_change.new_value) {
-            (Some(old), Some(new)) => {
-                println!(
-                    "        {}: \"{}\" → \"{}\"",
-                    prop_change.property, old, new
-                );
-            }
-            (Some(old), None) => {
-                println!("        {}: \"{}\" → (removed)", prop_change.property, old);
-            }
-            (None, Some(new)) => {
-                println!("        {}: (added) \"{}\"", prop_change.property, new);
-            }
-            (None, None) => {}
-        }
-    }
 }
